requirements.txt
absl-py==1.3.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrdict==2.0.1
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
click==8.1.3
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cupy==10.3.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.6.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fastrlock==0.8
filelock==3.8.0
flatbuffers==23.1.21
fonttools==4.32.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
hydra-core==1.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.21.3
imath==0.0.2
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.7.0
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.10
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyter-packaging==0.12.0
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.12.0
jupyterlab-widgets==3.0.3
keras==2.11.0
kiwisolver==1.4.2
libclang==15.0.6.1
loguru==0.6.0
magickwand==0.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nbclassic==0.4.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.8.6
nltk==3.7
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.23.4
oauthlib==3.2.0
omegaconf==2.2.3
open3d==0.15.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
OpenEXR==1.3.9
opt-einsum==3.3.0
oyaml==1.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
path.py==12.5.0
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.13.0
pynvrtc==9.2
PyOpenGL==3.1.6
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
pyquaternion==0.9.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
regex==2022.9.13
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.9
sapien @ file:///share/pengyang/sim2real_active_tactile/sapien-2.0.0.dev20211227-cp38-cp38-manylinux2014_x86_64.whl
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tabulate==0.8.10
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.1.1
tokenizer==3.4.2
tokenizers==0.13.2
tomlkit==0.11.4
torch==1.11.0
torch-tb-profiler==0.4.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transform3d==0.0.4
transforms3d==0.4.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.12
Wand==0.6.10
wandb==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
yacs==0.1.8
zipp==3.8.0
